// shared/generatedRankedPriorities.ts
export const USER_RANKS_BY_NORMALIZED_NAME = {
  "bagbazar": {
    "rank": 1,
    "priority": "S",
    "score": 4,
    "name": "Bagbazar Sarbojanin"
  },
// ...
} as const;